backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==9.5.0
python-dotenv==1.0.0
firebase-admin==6.2.0 